@@ -1,6 +1,6 @@
 module github.com/tagatac/goheif
 
-go 1.18
+go 1.22.3
 
 require (
 	github.com/golang/mock v1.6.0
