@@ -1,14 +1,17 @@
 package libde265
 
-//#cgo 386 amd64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_SSE4_1 -msse4.1 -Wno-constant-conversion
-//#cgo arm arm64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_ARM
 //#cgo CFLAGS: -I.
+//#cgo amd64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_SSE4_1 -msse4.1
+//#cgo arm64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_ARM
+//#cgo darwin,amd64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_SSE4_1 -msse4.1 -Wno-constant-conversion
+//#cgo darwin,amd64 CXXFLAGS: -Ilibde265 -I. -std=c++11 -DHAVE_ARM
 // #include <stdint.h>
 // #include <stdlib.h>
 // #include "libde265/de265.h"
 import "C"
 
 import (
+	"errors"
 	"fmt"
 	"image"
 	"unsafe"
@@ -31,7 +34,7 @@ func Fini() {
 func NewDecoder(opts ...Option) (*Decoder, error) {
 	p := C.de265_new_decoder()
 	if p == nil {
-		return nil, fmt.Errorf("Unable to create decoder")
+		return nil, errors.New("unable to create decoder")
 	}
 
 	dec := &Decoder{ctx: p, hasImage: false}
@@ -69,14 +72,14 @@ func (dec *Decoder) Push(data []byte) error {
 	totalSize := len(data)
 	for pos < totalSize {
 		if pos+4 > totalSize {
-			return fmt.Errorf("Invalid NAL data")
+			return errors.New("invalid NAL data")
 		}
 
 		nalSize := uint32(data[pos])<<24 | uint32(data[pos+1])<<16 | uint32(data[pos+2])<<8 | uint32(data[pos+3])
 		pos += 4
 
 		if pos+int(nalSize) > totalSize {
-			return fmt.Errorf("Invalid NAL size: %d", nalSize)
+			return fmt.Errorf("invalid NAL size: %d", nalSize)
 		}
 
 		C.de265_push_NAL(dec.ctx, unsafe.Pointer(&data[pos]), C.int(nalSize), C.de265_PTS(0), nil)
@@ -98,13 +101,13 @@ func (dec *Decoder) DecodeImage(data []byte) (image.Image, error) {
 	}
 
 	if ret := C.de265_flush_data(dec.ctx); ret != C.DE265_OK {
-		return nil, fmt.Errorf("flush_data error")
+		return nil, fmt.Errorf("flush_data error: %d", ret)
 	}
 
 	var more C.int = 1
 	for more != 0 {
 		if decerr := C.de265_decode(dec.ctx, &more); decerr != C.DE265_OK {
-			return nil, fmt.Errorf("decode error")
+			return nil, fmt.Errorf("decode error: %d", decerr)
 		}
 
 		for {
@@ -153,9 +156,14 @@ func (dec *Decoder) DecodeImage(data []byte) (image.Image, error) {
 				ycc.Cb = C.GoBytes(unsafe.Pointer(cb), C.int(cheight*cstride))
 				ycc.Cr = C.GoBytes(unsafe.Pointer(cr), C.int(cheight*cstride))
 			} else {
-				ycc.Y = (*[1 << 30]byte)(unsafe.Pointer(y))[:int(height)*int(ystride)]
-				ycc.Cb = (*[1 << 30]byte)(unsafe.Pointer(cb))[:int(cheight)*int(cstride)]
-				ycc.Cr = (*[1 << 30]byte)(unsafe.Pointer(cr))[:int(cheight)*int(cstride)]
+				// Calculate the exact sizes needed
+				ySize := int(height) * int(ystride)
+				cSize := int(cheight) * int(cstride)
+
+				// Create slices directly from pointers with exact sizes
+				ycc.Y = unsafe.Slice((*byte)(y), ySize)
+				ycc.Cb = unsafe.Slice((*byte)(cb), cSize)
+				ycc.Cr = unsafe.Slice((*byte)(cr), cSize)
 			}
 
 			//C.de265_release_next_picture(dec.ctx)
@@ -164,5 +172,5 @@ func (dec *Decoder) DecodeImage(data []byte) (image.Image, error) {
 		}
 	}
 
-	return nil, fmt.Errorf("No picture")
+	return nil, errors.New("no picture")
 }
