@@ -274,7 +274,8 @@ int  decode_CABAC_term_bit(CABAC_decoder* decoder)
 }
 
 
-
+// When we read past the end of the bitstream (which should only happen on faulty bitstreams),
+// we will eventually only return zeros.
 int  decode_CABAC_bypass(CABAC_decoder* decoder)
 {
   logtrace(LogCABAC,"[%3d] bypass r:%x v:%x\n",logcnt,decoder->range, decoder->value);
@@ -287,6 +288,10 @@ int  decode_CABAC_bypass(CABAC_decoder* decoder)
       if (decoder->bitstream_end > decoder->bitstream_curr) {
         decoder->bits_needed = -8;
         decoder->value |= *decoder->bitstream_curr++;
+      }
+      else {
+        // we read past the end of the bitstream, fill with 0
+        decoder->bits_needed = -8;
       }
     }
 
@@ -371,9 +376,9 @@ int  decode_CABAC_FL_bypass_parallel(CABAC_decoder* decoder, int nBits)
 }
 
 
-int  decode_CABAC_FL_bypass(CABAC_decoder* decoder, int nBits)
+uint32_t  decode_CABAC_FL_bypass(CABAC_decoder* decoder, int nBits)
 {
-  int value=0;
+  uint32_t value=0;
 
   if (likely(nBits<=8)) {
     if (nBits==0) {
@@ -485,6 +490,8 @@ void CABAC_encoder_bitstream::write_bits(uint32_t bits,int n)
   vlc_buffer |= bits;
   vlc_buffer_len += n;
 
+  // TODO: errors returned by append_byte() are ignored, resulting in a broken output.
+
   while (vlc_buffer_len>=8) {
     append_byte((vlc_buffer >> (vlc_buffer_len-8)) & 0xFF);
     vlc_buffer_len -= 8;
@@ -517,6 +524,8 @@ void CABAC_encoder::write_svlc(int value)
 
 void CABAC_encoder_bitstream::flush_VLC()
 {
+  // TODO: errors returned by append_byte() are ignored, resulting in a broken output.
+
   while (vlc_buffer_len>=8) {
     append_byte((vlc_buffer >> (vlc_buffer_len-8)) & 0xFF);
     vlc_buffer_len -= 8;
@@ -550,7 +559,7 @@ int  CABAC_encoder_bitstream::number_free_bits_in_byte() const
 }
 
 
-void CABAC_encoder_bitstream::check_size_and_resize(int nBytes)
+bool CABAC_encoder_bitstream::check_size_and_resize(int nBytes)
 {
   if (data_size+nBytes > data_capacity) { // 1 extra byte for stuffing
     if (data_capacity==0) {
@@ -559,14 +568,24 @@ void CABAC_encoder_bitstream::check_size_and_resize(int nBytes)
       data_capacity *= 2;
     }
 
-    data_mem = (uint8_t*)realloc(data_mem,data_capacity);
+    uint8_t* new_data_mem = (uint8_t*)realloc(data_mem,data_capacity);
+    if (new_data_mem) {
+      data_mem = new_data_mem;
+    }
+    else {
+      return false;
+    }
   }
+
+  return true;
 }
 
 
-void CABAC_encoder_bitstream::append_byte(int byte)
+bool CABAC_encoder_bitstream::append_byte(int byte)
 {
-  check_size_and_resize(2);
+  if (!check_size_and_resize(2)) {
+    return false;
+  }
 
   // --- emulation prevention ---
 
@@ -596,17 +615,23 @@ void CABAC_encoder_bitstream::append_byte(int byte)
   // write actual data byte
 
   data_mem[ data_size++ ] = byte;
+
+  return true;
 }
 
 
-void CABAC_encoder_bitstream::write_startcode()
+bool CABAC_encoder_bitstream::write_startcode()
 {
-  check_size_and_resize(3);
+  if (!check_size_and_resize(3)) {
+    return false;
+  }
 
   data_mem[ data_size+0 ] = 0;
   data_mem[ data_size+1 ] = 0;
   data_mem[ data_size+2 ] = 1;
   data_size+=3;
+
+  return true;
 }
 
 void CABAC_encoder_bitstream::init_CABAC()
@@ -621,6 +646,8 @@ void CABAC_encoder_bitstream::init_CABAC()
 
 void CABAC_encoder_bitstream::flush_CABAC()
 {
+  // TODO: errors returned by append_byte() are ignored, resulting in a broken output.
+
   if (low >> (32 - bits_left))
     {
       append_byte(buffered_byte + 1);
@@ -654,6 +681,8 @@ void CABAC_encoder_bitstream::flush_CABAC()
 
 void CABAC_encoder_bitstream::write_out()
 {
+  // TODO: errors returned by append_byte() are ignored, resulting in a broken output.
+  
   //logtrace(LogCABAC,"low = %08x (bits_left=%d)\n",low,bits_left);
   int leadByte = low >> (24 - bits_left);
   bits_left += 8;
