@@ -28,9 +28,9 @@
 #define LIBDE265_VERSION_H
 
 /* Numeric representation of the version */
-#define LIBDE265_NUMERIC_VERSION 0x01000300
+#define LIBDE265_NUMERIC_VERSION 0x01001500
 
 /* Version string */
-#define LIBDE265_VERSION "0.10.0"
+#define LIBDE265_VERSION "1.0.15"
 
 #endif
