@@ -217,7 +217,14 @@ void apply_sao_internal(de265_image* img, int xCtb,int yCtb,
           if (bandShift >= 8) {
             bandIdx = 0;
           } else {
-            bandIdx = bandTable[ in_img[xC+i+(yC+j)*in_stride]>>bandShift ];
+            int pixel = in_img[xC+i+(yC+j)*in_stride];
+
+            // Note: the input pixel value should never exceed the valid range, but it seems that it still does,
+            // maybe when there was a decoding error and the pixels have not been filled in correctly.
+            // Thus, we have to limit the pixel range to ensure that we have no illegal table access.
+            pixel = Clip3(0,maxPixelValue, pixel);
+
+            bandIdx = bandTable[ pixel>>bandShift ];
           }
 
           if (bandIdx>0) {
@@ -245,7 +252,14 @@ void apply_sao_internal(de265_image* img, int xCtb,int yCtb,
             if (bandShift >= 8) {
               bandIdx = 0;
             } else {
-              bandIdx = bandTable[ in_img[xC+i+(yC+j)*in_stride]>>bandShift ];
+              int pixel = in_img[xC+i+(yC+j)*in_stride];
+
+              // Note: the input pixel value should never exceed the valid range, but it seems that it still does,
+              // maybe when there was a decoding error and the pixels have not been filled in correctly.
+              // Thus, we have to limit the pixel range to ensure that we have no illegal table access.
+              pixel = Clip3(0,maxPixelValue, pixel);
+
+              bandIdx = bandTable[ pixel>>bandShift ];
             }
 
             if (bandIdx>0) {
